% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{fbds_cache_set}
\alias{fbds_cache_set}
\title{Definir o diretorio de cache do pacote}
\usage{
fbds_cache_set(path)
}
\arguments{
\item{path}{Caminho do diretorio a usar. Criado se nao existir.}
}
\value{
\code{path}, invisivelmente.
}
\description{
Definir o diretorio de cache do pacote
}
