% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{fbds_ufs}
\alias{fbds_ufs}
\title{Listar as 27 UFs com contagem de municipios}
\usage{
fbds_ufs()
}
\value{
Um tibble com \code{uf} e \code{n_municipios}.
}
\description{
Listar as 27 UFs com contagem de municipios
}
\examples{
fbds_ufs()
}
