% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{fbds_validate}
\alias{fbds_validate}
\title{Validar completude dos conjuntos de shapefile de um manifesto}
\usage{
fbds_validate(manifest)
}
\arguments{
\item{manifest}{Um manifesto de \code{\link[=fbds_fetch]{fbds_fetch()}} ou
\code{\link[=fbds_download_municipality]{fbds_download_municipality()}}.}
}
\value{
\code{manifest}, invisivelmente.
}
\description{
Agrupa os arquivos \code{.shp}/\code{.shx}/\code{.dbf}/\code{.prj} por nome-base e avisa
(classe \code{fbds_incomplete_shapefile}) para cada conjunto com algum
componente faltando.
}
\examples{
\dontrun{
m <- fbds_download_municipality("1100031", layers = "app")
fbds_validate(m)
}
}
