% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geofbds-package.R
\docType{package}
\name{geofbds-package}
\alias{geofbds}
\alias{geofbds-package}
\title{geofbds: Download and Read Municipal Data from Geo FBDS}
\description{
Discovers, downloads, and reads geospatial land-use, hydrography, and permanent preservation area (APP) data published by the Fundacao Brasileira para o Desenvolvimento Sustentavel (FBDS) for Brazilian municipalities. Provides a resolver from IBGE geocode, municipality name, or state (UF) to a frozen internal catalog, parallel resumable downloads via curl, and an auditable manifest of every transfer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kguidonimartins/geofbds}
  \item \url{https://kguidonimartins.github.io/geofbds/}
  \item Report bugs at \url{https://github.com/kguidonimartins/geofbds/issues}
}

}
\author{
\strong{Maintainer}: Karlo Guidoni \email{kguidonimartins@gmail.com}

Authors:
\itemize{
  \item Karlo Guidoni \email{kguidonimartins@gmail.com}
}

}
\keyword{internal}
